+++ packages/entities/entities-plugins/src/components/free-form/shared/Form.vue
@@ -30,8 +30,7 @@ import { FIELD_RENDERERS, provideFormShared } from './composables'
 import type { FormSchema, UnionFieldSchema } from '../../../types/plugins/form-schema'
 import Field from './Field.vue'
 import type { FormConfig, GlobalAction, RenderRules } from './types'
-import { filterByDependencies, sortFieldsByBundles, sortFieldsByFieldNames } from './utils'
-import { get } from 'lodash-es'
+import { sortFieldsByBundles, sortFieldsByFieldNames } from './utils'
 
 defineOptions({ name: 'SchemaForm' })
 
@@ -51,7 +50,7 @@ const emit = defineEmits<{
 
 const slots = useSlots()
 
-const { getSchema, formData, resetFormData, rootRenderRules } = provideFormShared({
+const { getSchema, formData, setValue, getValue, rootRenderRules } = provideFormShared({
   schema,
   propsData: computed(() => data as T),
   propsConfig: config as FormConfig,
@@ -63,14 +62,6 @@ const childFields = computed(() => {
   const { fields } = getSchema()
 
   let sortedFields = [...fields]
-
-  if (rootRenderRules.value?.dependencies) {
-    sortedFields = filterByDependencies(
-      sortedFields,
-      rootRenderRules.value.dependencies,
-      fieldName => get(formData, fieldName),
-    )
-  }
 
   if (fieldsOrder) {
     return sortFieldsByFieldNames(sortedFields, fieldsOrder)
@@ -85,7 +76,8 @@ const childFields = computed(() => {
 
 
 defineExpose({
-  getInnerData: () => toRaw(formData),
-  setInnerData: (data: T) => resetFormData(data),
+  getValue,
+  getRawValue: () => toRaw(formData),
+  setValue,
 })
 </script>