+++ packages/entities/entities-plugins/src/components/free-form/shared/RedisSelector.vue
@@ -1,6 +1,7 @@
 <template>
   <KCard
     v-if="useRedisPartial"
+    v-show="!hide"
     class="redis-config-card"
     data-testid="redis-config-card"
     :title="t('redis.title')"
@@ -17,11 +18,7 @@
         :description="t('redis.shared_configuration.description')"
         :label="t('redis.shared_configuration.label')"
         :selected-value="true"
-      >
-        <KBadge appearance="success">
-          {{ t('redis.shared_configuration.badge') }}
-        </KBadge>
-      </KRadio>
+      />
       <KRadio
         v-model="usePartial"
         card
@@ -81,6 +78,7 @@
   </KCard>
   <ObjectField
     v-else
+    v-show="!hide"
     v-bind="props"
     :fields-order="fieldsOrder"
     hide-required-asterisk
@@ -146,7 +144,7 @@ const partialsSaved = ref<PartialArray | undefined>()
 // initialize getter and setter for redis partial/redis fields
 const { value: partialValue } = useFormData<PartialArray | null | undefined>('$.partials')
 
-const { value: redisFieldsValue } = useField<Redis | undefined>(formRedisPath)
+const { value: redisFieldsValue, hide } = useField<Redis | undefined>(formRedisPath)
 
 const formConfig : KonnectBaseFormConfig | KongManagerBaseFormConfig = inject(FORMS_CONFIG)!